app/flaskProject/requirements.txt
Flask==2.3.3
pyserial==3.5
RPi.GPIO==0.7.1
mfrc522==0.0.7